+++ adjust-module-biz/src/main/java/com/hongguoyan/module/biz/controller/admin/nationalscore/vo/NationalScoreSaveReqVO.java
@@ -14,7 +14,7 @@ public class NationalScoreSaveReqVO {
 
     @Schema(description = "年份", requiredMode = Schema.RequiredMode.REQUIRED)
     @NotNull(message = "年份不能为空")
-    private Short year;
+    private Integer year;
 
     @Schema(description = "学位类型(0=不区分,1=专硕,2=学硕)", requiredMode = Schema.RequiredMode.REQUIRED, example = "1")
     @NotNull(message = "学位类型(0=不区分,1=专硕,2=学硕)不能为空")
